Abstract out helper functions to HomuraUtils.java
